requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
easyocr==1.7.2
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
ipykernel==6.30.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.24.3
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
Pygments==2.19.2
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tifffile==2025.5.10
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
